Bugfixes/refactoring Fixed edge case where 'y' being viewed as a consonant made some valid words invalid in basic title checker Moved title checkers and tests to respective packages
