+++ src/main/java/Main.java
@@ -3,8 +3,8 @@ import adapters.controllers.*;
 import adapters.presenters.*;
 import entities.LobbyManager;
 import entities.PlayerFactory;
-import entities.SuggestedTitleChecker;
-import entities.SuggestedTitleCheckerBasic;
+import entities.suggested_title_checkers.SuggestedTitleChecker;
+import entities.suggested_title_checkers.SuggestedTitleCheckerBasic;
 import entities.display_name_checkers.DisplayNameChecker;
 import entities.display_name_checkers.DisplayNameCheckerBasic;
 import entities.games.GameFactory;
@@ -16,7 +16,6 @@ import usecases.disconnecting.DcInteractor;
 import usecases.get_latest_stories.GlsInteractor;
 import usecases.get_most_liked_stories.GmlsInteractor;
 import usecases.join_public_lobby.JplInteractor;
-import usecases.like_story.LsGatewayStory;
 import usecases.like_story.LsInteractor;
 import usecases.pull_data.PdInteractor;
 import usecases.pull_game_ended.PgeInteractor;
